When manipulating large objects we must handle the oid as a Java long rather than an int so it works when the oid counter exceeds Integer.MAX_VALUE.  This doesn't remove the existing int API to avoid breaking code, but instead adds a parallel API that uses long.

+++ org/postgresql/jdbc2/AbstractJdbc2Blob.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2004-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/AbstractJdbc2Blob.java,v 1.10 2006/11/05 05:58:22 jurka Exp $
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/AbstractJdbc2Blob.java,v 1.11 2007/02/19 06:00:24 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -17,7 +17,7 @@ import org.postgresql.largeobject.LargeObject;
 public abstract class AbstractJdbc2Blob extends AbstractJdbc2BlobClob
 {
 
-    public AbstractJdbc2Blob(PGConnection conn, int oid) throws SQLException
+    public AbstractJdbc2Blob(PGConnection conn, long oid) throws SQLException
     {
         super(conn, oid);
     }
